input: focaltech_touch_mi: Update and handle aod_status
Field aod_status is currently unused because its value is always
zero. In stock ROM its value is updated by touchfeature HAL that
we do not use.
Set this field to 1 when display goes to low-power (doze) mode
and to 0 when display resumes. During doze mode (aod_status == 1)
when an user touches FOD and interrupt is generated we have to
wake-up event to resume device.
Note that we need to avoid to reset touchscreen device during
resume callback otherwise an user would be forced to touch
FOD twice.

Signed-off-by: Ivan Vecera <ivan@cera.cz>
Change-Id: I7d298d3a186a4647d69d27297d5ea66f093b3b2e